[FIX] Buttongroupattribute & force draw as unityobject instead of gameobject
