+++ Assets/GUIUtils/Editor/GUI/Drawables/Composite/BaseVerticalGroupDrawable.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 1e7e33bee16d438d95146140ae0d914f
+timeCreated: 1680594407
