+++ Assets/GUIUtils/Editor/GUI/Drawables/Composite/GroupingHelper.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using Rhinox.Lightspeed;
 using Sirenix.OdinInspector;
+using UnityEngine;
 
 namespace Rhinox.GUIUtils.Editor
 {
@@ -11,13 +12,22 @@ namespace Rhinox.GUIUtils.Editor
         
         public static GroupedDrawable CreateFrom(PropertyGroupAttribute groupingAttr, GroupedDrawable parent = null)
         {
-            if (groupingAttr is TitleGroupAttribute)
-                return new TitleGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
-            
-            if (groupingAttr is HorizontalGroupAttribute || groupingAttr is ButtonGroupAttribute)
-                return new HorizontalGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+            switch (groupingAttr)
+            {
+                case ButtonGroupAttribute buttonGroupAttribute:
+                    return new ButtonGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+                case HorizontalGroupAttribute horizontalGroupAttribute:
+                    return new HorizontalGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+                case TitleGroupAttribute titleGroupAttribute:
+                    return new TitleGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+                case VerticalGroupAttribute verticalGroupAttribute:
+                    return new VerticalGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+                default:
+                    Debug.LogError($"Unimplemented Group type: {groupingAttr.GetType().Name}");
+                    break;
+            }
 
-            return new VerticalGroupDrawable(parent, groupingAttr.GroupID, groupingAttr.Order);
+            return null;
         }
 
         public static string[] SplitIntoParts(string groupId)
