+++ Assets/GUIUtils/Editor/GUI/Drawables/Composite/PropertyGroupDrawable.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 34524f51e858491a9ed9dac0bfd24bca
+timeCreated: 1680593084
